# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Material Formulation API
  description: API for managing and querying material formulations for the POC.
  version: 1.1.0
servers:
  - url: http://localhost:4566/restapis/{api_id}/prod/_user_request_/
    description: LocalStack REST API (Replace {api_id} with the ID from `aws apigateway get-rest-apis`)

paths:
  /api/v1/materials:
    get:
      summary: List all materials
      description: Returns a list of all material formulations stored in the system.
      parameters:
        - name: limit
          in: query
          description: The number of items to return (default 10).
          schema:
            type: integer
            default: 10
        - name: cursor
          in: query
          description: The pagination cursor (Base64 encoded) for the next page.
          schema:
            type: string
      x-amazon-apigateway-integration:
        uri: arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:000000000000:function:PoC-Materials/invocations
        passthroughBehavior: when_no_match
        httpMethod: POST
        type: aws_proxy
      responses:
        200:
          description: A paged list of materials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponseMaterialFormulation'
        500:
          $ref: '#/components/responses/ServerError'
    post:
      summary: Create a new material formulation
      description: Stores a new material formulation in the DynamoDB table.
      x-amazon-apigateway-integration:
        uri: arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:000000000000:function:PoC-Materials/invocations
        payloadFormatVersion: "2.0"
        type: aws_proxy
        httpMethod: POST
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialFormulation'
      responses:
        201:
          description: Material created successfully.
        400:
          description: Invalid input.
        500:
          $ref: '#/components/responses/ServerError'

  /api/v1/materials/count:
    get:
      summary: Get total material count
      description: Returns the total number of material formulations in the database.
      x-amazon-apigateway-integration:
        uri: arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:000000000000:function:PoC-Materials/invocations
        passthroughBehavior: when_no_match
        httpMethod: POST
        type: aws_proxy
      responses:
        200:
          description: Total count of materials.
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 42
        500:
          $ref: '#/components/responses/ServerError'

  /api/v1/materials/{material_id}:
    get:
      summary: Get a material by ID
      parameters:
        - name: material_id
          in: path
          required: true
          schema:
            type: string
      x-amazon-apigateway-integration:
        uri: arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:000000000000:function:PoC-Materials/invocations
        payloadFormatVersion: "2.0"
        type: aws_proxy
        httpMethod: POST
      responses:
        200:
          description: Material formulation details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialFormulation'
        404:
          description: Material not found.
        500:
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete a material by ID (hard delete)
      parameters:
        - name: material_id
          in: path
          required: true
          schema:
            type: string
      x-amazon-apigateway-integration:
        uri: arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:000000000000:function:PoC-Materials/invocations
        payloadFormatVersion: "2.0"
        type: aws_proxy
        httpMethod: POST
      responses:
        204:
          description: Material deleted successfully.
        404:
          description: Material not found.
        500:
          $ref: '#/components/responses/ServerError'

  /api/v1/populator/jobs:
    post:
      summary: Populates the database with random data
      description: Triggers an asynchronous job to generate and insert random material formulations.
      x-amazon-apigateway-integration:
        uri: arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:000000000000:function:PoC-Populator/invocations
        passthroughBehavior: when_no_match
        httpMethod: POST
        type: aws_proxy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PopulationRequest'
      responses:
        202:
          description: Population job accepted.
        400:
          description: Invalid request or count exceeds limit.
        500:
          $ref: '#/components/responses/ServerError'

  /api/v1/costing/prices:
    post:
      summary: Upsert a component price
      description: Creates or updates the price for a component in the costing catalog.
      x-amazon-apigateway-integration:
        uri: arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:000000000000:function:PoC-Costing/invocations
        passthroughBehavior: when_no_match
        httpMethod: POST
        type: aws_proxy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComponentPriceRequest'
      responses:
        200:
          description: Price updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComponentPriceResponse'
        400:
          $ref: '#/components/responses/ValidationError'
        500:
          $ref: '#/components/responses/ServerError'

  /api/v1/costing/estimations:
    post:
      summary: Calculate material cost
      description: Calculates the total cost and breakdown for a material formulation based on component prices.
      x-amazon-apigateway-integration:
        uri: arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:000000000000:function:PoC-Costing/invocations
        payloadFormatVersion: "2.0"
        type: aws_proxy
        httpMethod: POST
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostCalculationRequest'
      responses:
        200:
          description: Cost calculation successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostCalculationResponse'
        400:
          $ref: '#/components/responses/ValidationError'
        500:
          $ref: '#/components/responses/ServerError'

  /api/v1/costing/estimations/batch:
    get:
      summary: Calculate costs for all materials
      description: Retrieves all materials and calculates their costs based on specific component prices.
      x-amazon-apigateway-integration:
        uri: arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:000000000000:function:PoC-Costing/invocations
        passthroughBehavior: when_no_match
        httpMethod: POST
        type: aws_proxy
      responses:
        200:
          description: Batch cost calculation successful.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CostCalculationResponse'
        500:
          $ref: '#/components/responses/ServerError'


components:
  schemas:
    PopulationRequest:
      type: object
      properties:
        target:
          type: string
          enum: [materials, prices, ensure-prices]
          default: materials
          description: The target dataset to populate. 'materials' generates formulations; 'ensure-prices' ensures all components have prices.
        count:
          type: integer
          maximum: 100000
          example: 1000
        minComponents:
          type: integer
          minimum: 1
          maximum: 1000
          default: 20
          description: Minimum number of components per material (for target='materials').
        maxComponents:
          type: integer
          minimum: 1
          maximum: 1000
          default: 20
          description: Maximum number of components per material (for target='materials').
    PagedResponseMaterialFormulation:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/MaterialFormulation'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'

    PaginationMeta:
      type: object
      properties:
        limit:
          type: integer
          description: The number of items per page.
        count:
          type: integer
          description: The number of items in the current page.
        nextCursor:
          type: string
          nullable: true
          description: The cursor to fetch the next page (Base64 encoded).

    Link:
      type: object
      properties:
        rel:
          type: string
          description: The relationship of the link (e.g., self, next).
        href:
          type: string
          description: The absolute URL of the link.
        method:
          type: string
          description: The HTTP method for the link (e.g., GET).

    MaterialFormulation:
      type: object
      required:
        - material_id
        - name
      properties:
        material_id:
          type: string
          example: composite-x-100
        name:
          type: string
          example: High-Strength Polymer Blend
        density:
          type: object
          properties:
            value:
              type: number
            unit:
              type: string
        formulation:
          type: array
          items:
            type: object
            properties:
              component:
                type: string
              percentage:
                type: number
              type:
                type: string
        properties:
          type: object
          additionalProperties:
            type: string

    ComponentPriceRequest:
      type: object
      required:
        - component_name
        - unit_price
        - unit
        - currency
      properties:
        component_name:
          type: string
          example: Polycarbonate
        unit_price:
          type: number
          format: decimal
          example: 12.50
        unit:
          type: string
          example: kg
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: USD

    ComponentPriceResponse:
      type: object
      properties:
        component_name:
          type: string
        unit_price:
          type: number
          format: decimal
        unit:
          type: string
        currency:
          type: string
        updated_at:
          type: string
          format: date-time

    CostCalculationRequest:
      type: object
      required:
        - material_id
        - formulation
      properties:
        material_id:
          type: string
          example: composite-x-100
        formulation:
          type: array
          items:
            type: object
            required:
              - component
              - percentage
            properties:
              component:
                type: string
                example: Polycarbonate
              percentage:
                type: number
                example: 60.0
        desired_margin_percent:
          type: number
          format: decimal
          example: 25.0

    CostCalculationResponse:
      type: object
      properties:
        material_id:
          type: string
        total_cost:
          type: number
          format: decimal
        currency:
          type: string
        breakdown:
          type: array
          items:
            type: object
            properties:
              component:
                type: string
              percentage:
                type: number
              unit_price:
                type: number
                format: decimal
              contribution_cost:
                type: number
                format: decimal
        margin:
          type: object
          nullable: true
          properties:
            desired_margin_percent:
              type: number
              format: decimal
            suggested_selling_price:
              type: number
              format: decimal
            estimated_gross_profit:
              type: number
              format: decimal

  responses:
    ServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              detail:
                type: string

    ValidationError:
      description: Validation error (RFC 7807 ProblemDetails).
      content:
        application/problem+json:
          schema:
            type: object
            properties:
              type:
                type: string
                example: about:blank
              title:
                type: string
                example: Validation failed
              status:
                type: integer
                example: 400
              detail:
                type: string
                example: Formulation percentages must sum to 100%